chapter_001/src/test/java/ru/job4j/array/FindLoopTest.java
package ru.job4j.array;


import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class FindLoopTest {
    @Test
    public void whenArrayHasLengh5Then0() {
        FindLoop find = new FindLoop();
        int[] input = new int[] {5, 10, 3};
        int value = 5;
        int result = find.indexOf(input, value);
        int expect = 0;
        assertThat(result, is(expect));
    }

    @Test
    public void whenArrayHasValue3AtPosition2Then2() {
        FindLoop find = new FindLoop();
        int[] input = new int[] {1, 2, 3, 4};
        int value = 3;
        int result = find.indexOf(input, value);
        int expect = 2;
        assertThat(result, is(expect));
    }

    @Test
    public void whenSearchForNotExistentElementThenMinusOne() {
        FindLoop find = new FindLoop();
        int[] input = new int[] {1, 2, 3, 4, 5, 6};
        int value = 9;
        int result = find.indexOf(input, value);
        int expect = -1;
        assertThat(result, is(expect));
    }

}